Export snippet execution code for dependent packages.
