+++ .gitignore
@@ -5,6 +5,7 @@ yarn-error.log
 local.log
 .DS_Store
 dist/
+scripts/**/*.js
 bazel-out/
 .idea/
 *.tgz
