+++ package.json
@@ -57,6 +57,8 @@
     "build": "./scripts/enumerate-tests.js && tsc",
     "build:bazel": "bazel build //...",
     "build-npm": "./scripts/build-npm.sh",
+    "build-test-snippets": "yarn tsc --project ./scripts/test_snippets/tsconfig.json",
+    "commander": "2.20.0",
     "link-local": "yalc link",
     "publish-local": "rimraf dist/ && yarn build && rollup -c && yalc push",
     "lint": "tslint -p . -t verbose",
